+++ BriefingRoomGUI/BriefingRoomGUI.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>GUI_API</RootNamespace>
+    <RootNamespace>BriefingRoomGUI</RootNamespace>
   </PropertyGroup>
 
     <ItemGroup>
