+++ BriefingRoomGUI/Controllers/GeneratorController.cs
