+++ BriefingRoomGUI/Pages/Error.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model GUI_API.Pages.ErrorModel
+@model BriefingRoomGUI.Pages.ErrorModel
 
 <!DOCTYPE html>
 <html>
