+++ BriefingRoomGUI/Pages/_Host.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@namespace GUI_API.Pages
+@namespace BriefingRoomGUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
     Layout = null;
@@ -14,7 +14,7 @@
     <base href="~/" />
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" crossorigin="anonymous">
     <link href="css/site.css" rel="stylesheet" />
-    <link href="GUI-API.styles.css" rel="stylesheet" />
+    <link href="BriefingRoomGUI.styles.css" rel="stylesheet" />
 </head>
 <body>
     <component type="typeof(App)" render-mode="ServerPrerendered" />
