+++ BriefingRoomGUI/Shared/MainLayout.razor
