+++ BriefingRoomGUI/Shared/MainLayout.razor.css
