+++ BriefingRoomGUI/Shared/NavMenu.razor
