+++ BriefingRoomGUI/Shared/NavMenu.razor.css
