+++ BriefingRoomGUI/Startup.cs
@@ -10,9 +10,9 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using GUI_API.Data;
+using BriefingRoomGUI.Data;
 
-namespace GUI_API
+namespace BriefingRoomGUI
 {
     public class Startup
     {
