+++ BriefingRoomGUI/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using GUI_API
-@using GUI_API.Shared
+@using BriefingRoomGUI
+@using BriefingRoomGUI.Shared
