+++ BriefingRoomGUI/wwwroot/css/open-iconic/font/fonts/open-iconic.eot
