+++ BriefingRoomGUI/wwwroot/js/download.js
