+++ briefing-room-for-dcs.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.6.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GUI-API", "GUI-API\GUI-API.csproj", "{654184B9-8AE6-4233-B38F-DFF5DCA9664E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BriefingRoomGUI", "BriefingRoomGUI\BriefingRoomGUI.csproj", "{654184B9-8AE6-4233-B38F-DFF5DCA9664E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BriefingRoom", "Source\BriefingRoom.csproj", "{62C3A919-BE9A-40AD-A867-F79D40CB46FB}"
 EndProject
